refactor(api): cleaning up the api directory now that core is split out
