+++ packages/api/src/codegen/language.ts
@@ -1,7 +1,7 @@
 import type Storage from '../storage';
 import type Oas from 'oas';
 
-import { PACKAGE_NAME, PACKAGE_VERSION } from '../../packageInfo';
+import { PACKAGE_NAME, PACKAGE_VERSION } from '../packageInfo';
 
 export interface InstallerOptions {
   /**
