+++ packages/api/src/codegen/languages/typescript/util.ts
