+++ packages/api/src/commands/install.ts
