+++ packages/api/src/lib/prompt.ts
