+++ packages/api/src/storage.ts
@@ -7,9 +7,8 @@ import makeDir from 'make-dir';
 import ssri from 'ssri';
 import validateNPMPackageName from 'validate-npm-package-name';
 
-import { PACKAGE_VERSION } from '../packageInfo';
-
 import Fetcher from './fetcher';
+import { PACKAGE_VERSION } from './packageInfo';
 
 export default class Storage {
   static dir: string;
