+++ packages/api/test/codegen/languages/typescript/utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 
-import { docblockEscape, generateTypeName, wordWrap } from '../../../../../src/cli/codegen/languages/typescript/util';
+import { docblockEscape, generateTypeName, wordWrap } from '../../../../src/codegen/languages/typescript/util';
 
 describe('ts codegen utils', () => {
   describe('#docblockEscape', () => {
